chore: changes label extention strategy
- our previous approach had a problem and a premature optimization:

- problem: because PyTorch expects a regular sized batch, we would
  have a problem whenever we batched two or more sentences with
  different lengths. Of course the tokenizer handles this for the input
  id, but not for the label.

- premature optimization: bert ignores all subword tokens when doing
  token level classification, so there is no need to extend labels for
  this purpose. Further experimentation will be performed to test this
  hypothesis properly, but there is no need to implement this now

- this commit, therefore, removes the subword label extension in favor
of adding a padding strategy to the labels, enforcing that all entries
in a batch are of same shape. A '-1' label is added for padding. This
choice is arbitrary because BERT simply ignores padding tokens in
classification, because of the attention mask. When calculating loss,
we'll need to remove the extra padding in the label. this can be easily
done with a find function.

Signed-off-by: Caio Andrade <caiotdandrade@gmail.com>