fix: changes dummy label from -1 to 0
- In order to extend the labels to cover the padding tokens
(which would not be classified due to the attention mask, but pytorch
asks for a regular shape dataset), a -1 label was used. This caused
a out of bounds exception. Because this label will just be ignored,
i've used 0 as this denotes a token without semantical meaning. the
exception was therefore fixed

Signed-off-by: Caio Andrade <caiotdandrade@gmail.com>